Add network to card client and request factory
