+++ Card/src/main/java/com/paypal/android/card/network/CardAPIClient.kt
@@ -2,29 +2,28 @@ package com.paypal.android.card.network
 
 import com.paypal.android.card.Card
 import com.paypal.android.card.CardResult
-import com.paypal.android.core.ConfirmPaymentSourceRequestFactory
+import com.paypal.android.core.Environment
 import com.paypal.android.core.Http
 import com.paypal.android.core.HttpRequest
-import com.paypal.android.core.PaymentsConfiguration
 import java.net.URL
 
-class CardAPIClient(private val config: PaymentsConfiguration) {
-    private val paymentSourceRequestFactory = ConfirmPaymentSourceRequestFactory(config)
+class CardAPIClient {
     private val http = Http()
+    private val baseUrl = Environment.SANDBOX.url
 
     suspend fun confirmPaymentSource(orderId: String, card: Card): CardResult {
-        val request = paymentSourceRequestFactory.create(orderId, card.toJson())
-        val clientToken = fetchClientToken(config)
+        val request = ConfirmPaymentSourceRequestFactory().create(orderId, card)
+        val clientToken = fetchClientToken()
         request.headers["Authorization"] = "Basic $clientToken"
 
         val response = http.send(request)
-        //response to CardResult
+            //response to CardResult
         return CardResult(response, null)
     }
 
-    private suspend fun fetchClientToken(config: PaymentsConfiguration): String {
+    private suspend fun fetchClientToken(): String {
         // fetch lsat
-        val fetchClientTokenUrl = URL("${config.environment.url}/v1/oauth2/token")
+        val fetchClientTokenUrl = URL("${baseUrl}/v1/oauth2/token")
         val body = """
             {
                 "grant_type": "client_credentials",
